requirements.txt
Django==4.2.2
psycopg2==2.9.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
django-modeltranslation==0.18.9
django-cors-headers==3.14.0
django-axes==5.41.1
django-ckeditor==6.5.1
django-filter==23.2
django-registration==3.3
django_bleach==3.0.1
django-storages==1.13.2
boto3==1.26.157
Pillow==9.5.0
django-allauth==0.50.0
dj-rest-auth==2.2.5
gunicorn==20.1.0
psycopg2-binary==2.9.6
django-heroku==0.3.1

